+++ internal/services/appservice/helpers/web_app_schema.go
@@ -2,13 +2,13 @@ package helpers
 
 import (
 	"fmt"
+	"github.com/hashicorp/go-azure-helpers/lang/pointer"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/services/web/mgmt/2021-02-01/web" // nolint: staticcheck
 	"github.com/Azure/go-autorest/autorest/date"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/features"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
 	apimValidate "github.com/hashicorp/terraform-provider-azurerm/internal/services/apimanagement/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
@@ -846,553 +846,6 @@ func SiteConfigSchemaLinuxComputed() *pluginsdk.Schema {
 
 				"vnet_route_all_enabled": {
 					Type:     pluginsdk.TypeBool,
-					Computed: true,
-				},
-			},
-		},
-	}
-}
-
-type ApplicationStackWindows struct {
-	NetFrameworkVersion     string `tfschema:"dotnet_version"`
-	PhpVersion              string `tfschema:"php_version"`
-	JavaVersion             string `tfschema:"java_version"`
-	PythonVersion           string `tfschema:"python_version"`
-	NodeVersion             string `tfschema:"node_version"`
-	JavaContainer           string `tfschema:"java_container"`
-	JavaContainerVersion    string `tfschema:"java_container_version"`
-	DockerContainerName     string `tfschema:"docker_container_name"`
-	DockerContainerRegistry string `tfschema:"docker_container_registry"`
-	DockerContainerTag      string `tfschema:"docker_container_tag"`
-	CurrentStack            string `tfschema:"current_stack"`
-}
-
-// Version information for the below validations was taken in part from - https://github.com/Azure/app-service-linux-docs/tree/master/Runtime_Support
-// There is no 3.0 version of .netFramework, setting the property to this value causing app to be broke.
-func windowsApplicationStackSchema() *pluginsdk.Schema {
-	return &pluginsdk.Schema{
-		Type:     pluginsdk.TypeList,
-		Optional: true,
-		Computed: true,
-		MaxItems: 1,
-		Elem: &pluginsdk.Resource{
-			Schema: map[string]*pluginsdk.Schema{
-				"dotnet_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: func() pluginsdk.SchemaValidateFunc {
-						if !features.FourPointOh() {
-							return validation.StringInSlice([]string{
-								"v2.0",
-								"v3.0",
-								"core3.1",
-								"v4.0",
-								"v5.0",
-								"v6.0",
-								"v7.0"}, false)
-						}
-						return validation.StringInSlice([]string{
-							"v2.0",
-							"core3.1",
-							"v4.0",
-							"v5.0",
-							"v6.0",
-							"v7.0",
-						}, false)
-					}(),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-				},
-
-				"php_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"7.4",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-				},
-
-				"python_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"3.4.0", // Note: The portal only lists `3.6`, however, the service only uses the value `3.4.0`. Anything else is silently discarded
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-				},
-
-				"node_version": { // Discarded by service if JavaVersion is specified
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"12-LTS",
-						"14-LTS",
-						"16-LTS",
-						"18-LTS",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.java_version",
-					},
-				},
-
-				"java_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"1.8",
-						"11",
-						"17",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-				},
-
-				"java_container": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"JAVA",
-						"JETTY",
-						"TOMCAT",
-					}, false),
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.java_container_version",
-					},
-				},
-
-				"java_container_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.java_container",
-					},
-				},
-
-				"docker_container_name": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty,
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-					},
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.docker_container_tag",
-					},
-				},
-
-				"docker_container_registry": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty,
-				},
-
-				"docker_container_tag": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty,
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.docker_container_name",
-					},
-				},
-
-				"current_stack": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"dotnet",
-						"dotnetcore",
-						"node",
-						"python",
-						"php",
-						"java",
-					}, false),
-				},
-			},
-		},
-	}
-}
-
-func windowsApplicationStackSchemaComputed() *pluginsdk.Schema {
-	return &pluginsdk.Schema{
-		Type:     pluginsdk.TypeList,
-		Computed: true,
-		Elem: &pluginsdk.Resource{
-			Schema: map[string]*pluginsdk.Schema{
-				"dotnet_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"php_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"python_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"node_version": { // Discarded by service if JavaVersion is specified
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_container": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_container_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"docker_container_name": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"docker_container_registry": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"docker_container_tag": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"current_stack": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-			},
-		},
-	}
-}
-
-type ApplicationStackLinux struct {
-	NetFrameworkVersion string `tfschema:"dotnet_version"`
-	PhpVersion          string `tfschema:"php_version"`
-	PythonVersion       string `tfschema:"python_version"` // Linux Only?
-	NodeVersion         string `tfschema:"node_version"`
-	JavaVersion         string `tfschema:"java_version"`
-	JavaServer          string `tfschema:"java_server"`
-	JavaServerVersion   string `tfschema:"java_server_version"`
-	DockerImageTag      string `tfschema:"docker_image_tag"`
-	DockerImage         string `tfschema:"docker_image"`
-	RubyVersion         string `tfschema:"ruby_version"`
-}
-
-// version information in the validation here was taken mostly from - `az webapp list-runtimes --linux`
-func linuxApplicationStackSchema() *pluginsdk.Schema {
-	return &pluginsdk.Schema{
-		Type:     pluginsdk.TypeList,
-		Optional: true,
-		Computed: true,
-		MaxItems: 1,
-		Elem: &pluginsdk.Resource{
-			Schema: map[string]*pluginsdk.Schema{
-				"dotnet_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"3.1",
-						"5.0",
-						"6.0",
-						"7.0",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.ruby_version",
-						"site_config.0.application_stack.0.java_version",
-					},
-				},
-
-				"php_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"7.4",
-						"8.0",
-						"8.1",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.ruby_version",
-						"site_config.0.application_stack.0.java_version",
-					},
-				},
-
-				"python_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"3.7",
-						"3.8",
-						"3.9",
-						"3.10",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-				},
-
-				"node_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"12-lts",
-						"14-lts",
-						"16-lts",
-						"18-lts",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-						"site_config.0.application_stack.0.java_version",
-					},
-				},
-
-				"ruby_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"2.6",
-						"2.7",
-					}, false),
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.java_version",
-					},
-				},
-
-				"java_version": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty, // There a significant number of variables here, and the versions are not uniformly formatted.
-					// TODO - Needs notes in the docs for this to help users navigate the inconsistencies in the service. e.g. jre8 va java8 etc
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					ConflictsWith: []string{
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-				},
-
-				"java_server": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-					ValidateFunc: validation.StringInSlice([]string{
-						"JAVA",
-						"TOMCAT",
-						"JBOSSEAP",
-					}, false),
-				},
-
-				"java_server_version": {
-					Type:     pluginsdk.TypeString,
-					Optional: true,
-				},
-
-				"docker_image": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty,
-					AtLeastOneOf: []string{
-						"site_config.0.application_stack.0.docker_image",
-						"site_config.0.application_stack.0.dotnet_version",
-						"site_config.0.application_stack.0.java_version",
-						"site_config.0.application_stack.0.node_version",
-						"site_config.0.application_stack.0.php_version",
-						"site_config.0.application_stack.0.python_version",
-						"site_config.0.application_stack.0.ruby_version",
-					},
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.docker_image_tag",
-					},
-				},
-
-				"docker_image_tag": {
-					Type:         pluginsdk.TypeString,
-					Optional:     true,
-					ValidateFunc: validation.StringIsNotEmpty,
-					RequiredWith: []string{
-						"site_config.0.application_stack.0.docker_image",
-					},
-				},
-			},
-		},
-	}
-}
-
-func linuxApplicationStackSchemaComputed() *pluginsdk.Schema {
-	return &pluginsdk.Schema{
-		Type:     pluginsdk.TypeList,
-		Computed: true,
-		Elem: &pluginsdk.Resource{
-			Schema: map[string]*pluginsdk.Schema{
-				"dotnet_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"php_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"python_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"node_version": { // Discarded by service if JavaVersion is specified
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"ruby_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_server": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"java_server_version": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"docker_image": {
-					Type:     pluginsdk.TypeString,
-					Computed: true,
-				},
-
-				"docker_image_tag": {
-					Type:     pluginsdk.TypeString,
 					Computed: true,
 				},
 			},
@@ -2927,12 +2380,9 @@ func ExpandSiteConfigWindows(siteConfig []SiteConfigWindows, existing *web.SiteC
 	if metadata.ResourceData.HasChange("site_config.0.application_stack") {
 		if len(winSiteConfig.ApplicationStack) == 1 {
 			winAppStack := winSiteConfig.ApplicationStack[0]
+			// TODO - only one of these should be non-nil?
 			expanded.NetFrameworkVersion = utils.String(winAppStack.NetFrameworkVersion)
-			if winAppStack.CurrentStack == "dotnetcore" {
-				expanded.NetFrameworkVersion = nil
-			}
 			expanded.PhpVersion = utils.String(winAppStack.PhpVersion)
-			expanded.NodeVersion = utils.String(winAppStack.NodeVersion)
 			expanded.PythonVersion = utils.String(winAppStack.PythonVersion)
 			expanded.JavaVersion = utils.String(winAppStack.JavaVersion)
 			expanded.JavaContainer = utils.String(winAppStack.JavaContainer)
@@ -3097,6 +2547,10 @@ func ExpandSiteConfigLinux(siteConfig []SiteConfigLinux, existing *web.SiteConfi
 				expanded.LinuxFxVersion = utils.String(fmt.Sprintf("DOTNETCORE|%s", linuxAppStack.NetFrameworkVersion))
 			}
 
+			if linuxAppStack.GoVersion != "" {
+				expanded.LinuxFxVersion = pointer.To(fmt.Sprintf("GO|%s", linuxAppStack.GoVersion))
+			}
+
 			if linuxAppStack.PhpVersion != "" {
 				expanded.LinuxFxVersion = utils.String(fmt.Sprintf("PHP|%s", linuxAppStack.PhpVersion))
 			}
@@ -3114,8 +2568,6 @@ func ExpandSiteConfigLinux(siteConfig []SiteConfigLinux, existing *web.SiteConfi
 			}
 
 			if linuxAppStack.JavaServer != "" {
-				// (@jackofallops) - Java has some special cases for Java SE when using specific versions of the runtime, resulting in this string
-				// being formatted in the form: `JAVA|u242` instead of the standard pattern of `JAVA|u242-java8` for example. This applies to jre8 and java11.
 				if linuxAppStack.JavaServer == "JAVA" && linuxAppStack.JavaServerVersion == "" {
 					expanded.LinuxFxVersion = utils.String(fmt.Sprintf("%s|%s", linuxAppStack.JavaServer, linuxAppStack.JavaVersion))
 				} else {
